Add converters and enable the slackbot dummy

+++ apps/stats/priv/repo/migrations/20160926173610_add_stats_table.exs
@@ -4,7 +4,7 @@ defmodule Stats.Repo.Migrations.AddStatsTable do
   def change do
     create table(:stats) do
       add :sensor_id, :string, size: 50, null: false
-      add :value, :integer, null: false
+      add :value, :float, null: false
 
       timestamps(updated_at: false)
     end
